feat: add bochs debugger
Adds a bochs config, a Makefile entry, and a wrapper around bochs magic
breakpoints to call from Rust code. This will help with much easier
debugging.

Future bochs configs should be stored in the bochs/ directory, and be
named bochs.(ARCH), where (ARCH) is the architecture. Similarly, future
bochs functions should be added to the bochs module.

Future-work: Add config attributes to ensure the bochs module is only
included on a bochs build.