feat(x86_64 idt): add basic x86_64 idt
Adds code to setup an Interrupt Descriptor Table for the x86_64
architecture, which allows registering interrupt and exception handlers.
There are a few VERY basic handlers that are included in this commit
that will surely need to be changed in the future as the kernel is
fleshed out more. This commit DOES NOT handle some double faults
correctly, and will triple fault if, say a page fault occurs on a kernel
guard page. This will be solved using a dedicated interrupt stack for
certain exceptions.