feat(alloc): add simple bump frame allocator
Added basic physical frame allocation, using a simple bump allocator,
which simply returns the frames in increasing order of memory.

A utils module was added, which currently only contains extra iterator
functions and types. I added an iterator called "Leftovers", which is
like map, but returns a two-tuple, where the first element is the
return, and the second is logically pushed back onto the list. This is
necessary because I cannot use `flat_map` (because I have no
vectors/memory allocation yet). The logic here and iterator could
probably be cleaned up and optimized substantially (though it may not be
worth the effort).

Future-work: Add more sophisticated frame allocators